Fix an issue regarding fetching apps.
