+++ config.py
@@ -29,8 +29,6 @@ class Config:
         if cls._instance is None:
             cls._instance = super(Config, cls).__new__(cls)
             cls._instance.load_environment_variables()
-            fetch_data = Thread(target=cls._instance.fetch_nodes_and_apps)
-            fetch_data.start()
         return cls._instance
 
     @staticmethod
@@ -126,11 +124,6 @@ class Config:
                 nodes[node_id]['stake'] = float(data['stake'])/(10**18)
         return nodes
     
-    
-    def fetch_apps(self) -> None:
-        """Fetchs the apps data."""
-        apps_data = Config.get_file_content(self.APPS_FILE)
-        self.APPS.update(apps_data)
         
     def fetch_nodes(self):
         """Fetchs the nodes data."""
@@ -147,18 +140,6 @@ class Config:
         self.NODE.update(nodes_data[self.ADDRESS])
         self.NODES.update(nodes_data)
     
-    def fetch_nodes_and_apps(self) -> None:
-        """Periodically fetches apps and nodes data."""
-        while True:
-            time.sleep(self.FETCH_APPS_AND_NODES_INTERVAL)
-            try:
-                self.fetch_apps()
-            except Exception:
-                zlogger.exception("An unexpected error occurred:")
-            try:
-                self.fetch_nodes()
-            except Exception:
-                zlogger.exception("An unexpected error occurred:")
 
     def register_operator(self, ecdsa_private_key, bls_key_pair) -> None:
         rpc_node = os.getenv('ZSEQUENCER_RPC_NODE')
