+++ node/routes.py
@@ -50,6 +50,8 @@ def put_transactions(app_name: str) -> Response:
     """Put a new batch into the database."""
     if not app_name:
         return error_response(ErrorCodes.INVALID_REQUEST, "app_name is required")
+    if app_name not in list(zconfig.APPS):
+        return error_response(ErrorCodes.INVALID_REQUEST, "invalid app_name.")
     data = request.data.decode('latin-1')
     zlogger.info(f"Transactions added. app: {app_name}, data length: {len(data)}.")
     zdb.init_batches(app_name, [data])
